unlock(achevement): :sparkles: Learned Traefik :technologist:
Such a cool reverse proxy! Config files are shorter than nginx and setup is simpler thanks mostly to the integration with Let's encrypt and nameservers APIs for TLS certs renewal via DNS challenge. The dashboard is also a great bonus. I have written a "getting started" guide on my own and will publish it on my own website when I spin it up.